Added decimal credit value support

+++ Graduate/Graduate.Droid/Fragments/EditDialogFragment.cs
@@ -408,7 +408,7 @@ namespace Graduate.Droid.Fragments
             IList<String> letterGrades = planner.getAllLetterGrades();
             try
             {
-                Convert.ToInt32(credits);
+                Convert.ToDouble(credits);
             }
             catch
             {
